+++ NM3_Driver/modem.h
@@ -13,32 +13,32 @@ class Modem
  public:
 	 Modem(wchar_t portnum);
 
-	 int Ping(unsigned Address);
+	 int64_t Ping(uint16_t Address);
 
-	 int SysTimeEnable(char& flag);
+	 int64_t SysTimeEnable(char& flag);
 
-	 int SysTimeGet(char& flag);
+	 int64_t SysTimeGet(char& flag);
 
-	 int SysTimeDisable(char& flag);
+	 int64_t SysTimeDisable(char& flag);
 
-	 int SysTimeClear(char& flag);
+	 int64_t SysTimeClear(char& flag);
 
-	 int Unicast(unsigned Address, char message[], unsigned messagelength, unsigned txtime = 0);
+	 int64_t Unicast(uint16_t Address, char message[], uint16_t messagelength, uint64_t txtime = 0);
 
-	 int UnicastWithAck(unsigned Address, char message[], unsigned messagelength, unsigned txtime = 0);
+	 int64_t UnicastWithAck(uint16_t Address, char message[], uint16_t messagelength, uint64_t txtime = 0);
 
-	 int Broadcast(char message[], unsigned messagelength, unsigned txtime = 0);
+	 int64_t Broadcast(char message[], uint16_t messagelength, uint64_t txtime = 0);
 
-	 int Probe(unsigned chirprepetitions, Chirp chirpinfo, unsigned txtime = 0);
+	 int64_t Probe(uint16_t chirprepetitions, Chirp chirpinfo, uint64_t txtime = 0);
 
-	 int UnicastListen(char *message);
+	 int64_t UnicastListen(char *message);
 
  private:
  
 	 wchar_t port_[9] = {L"\\\\.\\COM3"};	// default COM3
-	 const unsigned long baudrate_ = 9600;	// default 9600	
+	 const uint16_t baudrate_ = 9600;	// default 9600	
 
-	 int ConfigureSerial(HANDLE& hCom, DCB& dcb);
+	 int64_t ConfigureSerial(HANDLE& hCom, DCB& dcb);
 
-	 int SysTimeCommon(char& flag, char commandchar);
+	 int64_t SysTimeCommon(char& flag, char commandchar);
 };
